requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
joblib==1.3.2
MarkupSafe==2.1.3
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
yarl==1.9.2